requirements.txt
jinja2
PyYAML
sanic
nltk
